requirements.txt
# === Core ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# === Database ===
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# === Task Queue ===
celery[redis]==5.4.0
redis==5.2.1

# === LLM (ChatGPT Plus через веб-интерфейс + Playwright) ===
playwright==1.49.1

# === Financial Data ===
yfinance==0.2.51
feedparser==6.0.11
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0

# === Signal Engine ===
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
textblob==0.18.0
vaderSentiment==3.3.2

# === Telegram ===
python-telegram-bot==21.9

# === Utilities ===
httpx==0.28.1
loguru==0.7.3
tenacity==9.0.0
apscheduler==3.11.0

# === Testing ===
pytest==8.3.4
pytest-asyncio==0.25.0
